refactor(schemas): migrate Phase 37 endpoint schemas — 5 files, 10 classes (#6042)
Move local Pydantic class definitions from endpoint files into matching
domain schema modules. Renames applied to avoid name collisions.

- sequential_thinking_mcp.py: MCPTool → SequentialThinkingMCPTool (collides
  with schemas_code.MCPTool); SequentialThinkingRequest (with 9 helper
  methods preserved) → schemas_workflows.py
- state_tracking.py: StateChangeRequest → schemas_workflows.py;
  ExportRequest → StateTrackingExportRequest (collides with
  schemas_knowledge.ExportRequest)
- system_validation.py: ValidationRequest → SystemValidationRequestModel;
  ValidationResult → SystemValidationResultModel (consistency with existing
  ValidationResultModel naming) → schemas_workflows.py
- templates.py: TemplateExecutionRequest, TemplateValidationRequest →
  schemas_code.py
- web_research_settings.py: WebResearchSettings, ResearchPreferences →
  schemas_workflows.py

Endpoint files now import from the schemas modules; no behavior change.
